SLING-10299 : rename remove to delete as requested by suggested by bertrand
