+++ src/main/java/org/apache/sling/repoinit/parser/operations/OperationVisitor.java
@@ -38,8 +38,8 @@ public interface OperationVisitor {
     void visitAddGroupMembers(AddGroupMembers am);
     void visitRemoveGroupMembers(RemoveGroupMembers rm);
     void visitSetProperties(SetProperties sp);
-    default void visitRemoveAclPrincipals(RemoveAclPrincipals s) { throw new UnsupportedOperationException(); }
-    default void visitRemoveAclPaths(RemoveAclPaths s) { throw new UnsupportedOperationException(); }
-    default void visitRemoveAclPrincipalBased(RemoveAclPrincipalBased s) { throw new UnsupportedOperationException(); }
+    default void visitDeleteAclPrincipals(DeleteAclPrincipals s) { throw new UnsupportedOperationException(); }
+    default void visitDeleteAclPaths(DeleteAclPaths s) { throw new UnsupportedOperationException(); }
+    default void visitDeleteAclPrincipalBased(DeleteAclPrincipalBased s) { throw new UnsupportedOperationException(); }
 
 }
