+++ src/test/java/org/apache/sling/repoinit/parser/test/OperationToStringVisitor.java
@@ -26,6 +26,8 @@ import org.apache.sling.repoinit.parser.operations.CreateGroup;
 import org.apache.sling.repoinit.parser.operations.CreatePath;
 import org.apache.sling.repoinit.parser.operations.CreateServiceUser;
 import org.apache.sling.repoinit.parser.operations.CreateUser;
+import org.apache.sling.repoinit.parser.operations.DeleteAclPaths;
+import org.apache.sling.repoinit.parser.operations.DeleteAclPrincipalBased;
 import org.apache.sling.repoinit.parser.operations.DeleteGroup;
 import org.apache.sling.repoinit.parser.operations.DeleteServiceUser;
 import org.apache.sling.repoinit.parser.operations.DeleteUser;
@@ -34,9 +36,7 @@ import org.apache.sling.repoinit.parser.operations.RegisterNodetypes;
 import org.apache.sling.repoinit.parser.operations.OperationVisitor;
 import org.apache.sling.repoinit.parser.operations.RegisterNamespace;
 import org.apache.sling.repoinit.parser.operations.RegisterPrivilege;
-import org.apache.sling.repoinit.parser.operations.RemoveAclPaths;
-import org.apache.sling.repoinit.parser.operations.RemoveAclPrincipalBased;
-import org.apache.sling.repoinit.parser.operations.RemoveAclPrincipals;
+import org.apache.sling.repoinit.parser.operations.DeleteAclPrincipals;
 import org.apache.sling.repoinit.parser.operations.SetAclPaths;
 import org.apache.sling.repoinit.parser.operations.SetAclPrincipalBased;
 import org.apache.sling.repoinit.parser.operations.SetAclPrincipals;
@@ -130,7 +130,7 @@ class OperationToStringVisitor implements OperationVisitor {
     }
 
     @Override
-    public void visitRemoveAclPrincipals(RemoveAclPrincipals s) {
+    public void visitDeleteAclPrincipals(DeleteAclPrincipals s) {
         out.print(s.getClass().getSimpleName());
         out.print(" for");
         for (String p : s.getPrincipals()) {
@@ -141,7 +141,7 @@ class OperationToStringVisitor implements OperationVisitor {
     }
 
     @Override
-    public void visitRemoveAclPaths(RemoveAclPaths s) {
+    public void visitDeleteAclPaths(DeleteAclPaths s) {
         out.print(s.getClass().getSimpleName());
         out.print(" on");
         for (String p : s.getPaths()) {
@@ -153,7 +153,7 @@ class OperationToStringVisitor implements OperationVisitor {
     }
 
     @Override
-    public void visitRemoveAclPrincipalBased(RemoveAclPrincipalBased s) {
+    public void visitDeleteAclPrincipalBased(DeleteAclPrincipalBased s) {
         out.print(s.getClass().getSimpleName());
         out.print(" for");
         for (String p : s.getPrincipals()) {
