+++ src/test/resources/testcases/test-70.txt
@@ -1,13 +1,13 @@
-remove ACL for ana
+delete ACL for ana
 
-remove ACL for alice, aida
+delete ACL for alice, aida
 
-remove ACL on :repository, home(anni), functionNamesAreFree(aendu)
+delete ACL on :repository, home(anni), functionNamesAreFree(aendu)
 
-remove ACL on /, /var, /etc
+delete ACL on /, /var, /etc
 
-remove ACL on /content
+delete ACL on /content
 
-remove principal ACL for ada, amy
+delete principal ACL for ada, amy
 
-remove principal ACL for adi
+delete principal ACL for adi
