+++ src/test/resources/testcases/test-99.txt
@@ -137,10 +137,10 @@ set properties on authorizable(bob), authorizable(grpB)/nested
 end
 
 # Remove AC policies entirely (not just individual entries)
-remove ACL for ana
-remove ACL for alice, aida
-remove ACL on :repository, home(anni), functionNamesAreFree(aendu)
-remove ACL on /, /var, /etc
-remove ACL on /content
-remove principal ACL for ada, amy
-remove principal ACL for adi
+delete ACL for ana
+delete ACL for alice, aida
+delete ACL on :repository, home(anni), functionNamesAreFree(aendu)
+delete ACL on /, /var, /etc
+delete ACL on /content
+delete principal ACL for ada, amy
+delete principal ACL for adi
